libs/api/pyproject.toml
[project]
name = "hiagent-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "volcengine>=1.0.187",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "aiofiles>=23.1.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.sdist]
exclude = [
    "examples/",
]